Rename to tract_model and ort_model
