+++ src/api.rs
@@ -1,7 +1,7 @@
 use crate::{
     io::Query,
-    //model::{AppData, Model},
-    session::{AppData, Model},
+    tract_model::{AppData, Model},
+    //ort_model::{AppData, Model},
 };
 use actix_web::web;
 use actix_web::web::Data;
